added auth api and admin check

+++ src/routes/blog.js
@@ -5,15 +5,19 @@ const blogService = require("../service/blog");
 const router = express.Router();
 
 
-router.post("/blogs/push", auth, async (req, res) => {
+router.post("/blogs/push", auth, async (req, res, next) => {
     try {
-        console.log(req.body.title);
+        if (req.user.role !== "CONTENT-WRITER") {
+            const error = new Error("Please Authenticate as CONTENT-WRITER")
+            error.status = 401;
+            throw error;
+        }
         req.body.owner = req.user._id;
         const blog = new Blog(req.body);
         const insertedData = await blogService.push(blog);
         res.json(insertedData);
     } catch (error) {
-        res.status(500).send(error);
+        next(error);
     }
 });
 // for admin
@@ -38,6 +42,11 @@ router.get("/blogs/showactiveposts", async (req, res, next) => {
 
 router.post("/blogs/approve", auth, async (req, res, next) => {
     try {
+        if (req.user.role !== "ADMIN") {
+            const error = new Error("Please Authenticate as ADMIN")
+            error.status = 401;
+            throw error;
+        }
         const data = await blogService.approve(req.body.blogid);
         res.json(data);
     } catch (error) {
@@ -47,6 +56,11 @@ router.post("/blogs/approve", auth, async (req, res, next) => {
 
 router.post("/blogs/delete", auth, async (req, res, next) => {
     try {
+        if (req.user.role !== "ADMIN") {
+            const error = new Error("Please Authenticate as ADMIN")
+            error.status = 401;
+            throw error;
+        }
         const data = await blogService.delete(req.body.blogid);
         res.json(data);
     } catch (error) {
